# public/openapi.yaml
openapi: "3.1.0"
servers:
  - url: https://ccapdev-mco-pq20.onrender.com/api
info:
  version: "0.1.0"
  title: "ArchersLab API"
  description: CCAPDEV MCO
  license:
    name: MIT
    identifier: MIT
security:
  - Jwt: []
  - {}
paths:
...
  /user:
    post:
      summary: Create a user
      operationId: createUser
      tags:
        - user
      responses:
        "201":
          description: The user created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserId"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Exists"
        "500":
          $ref: "#/components/responses/Unexpected"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UserLogin"
                - type: object
                  required:
                    - name
                  properties:
                    name:
                      $ref: "#/components/schemas/UserName"
    get:
      summary: Get current user ID
      operationId: readCurrentUser
      tags:
        - user
      security:
        - Jwt: []
      responses:
        "200":
          description: Current user ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserId"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
  /user/{id}:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      summary: Read a user
      operationId: readUser
      tags:
        - user
      responses:
        "200":
          description: The user corresponding to the provided `id`
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/UserDetails"
                  - type: object
                    required:
                      - id
                    properties:
                      id:
                        $ref: "#/components/schemas/UserId"
                      admin:
                        type: boolean
                      editable:
                        $ref: "#/components/schemas/Editable"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
    put:
      summary: Update a user
      operationId: updateUser
      tags:
        - user
      security:
        - Jwt: []
      responses:
        "204":
          description: Update the user corresponding to the provided `id`
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/UserDetails"
                - type: object
                  properties:
                    password:
                      $ref: "#/components/schemas/UserPassword"
    delete:
      summary: Delete a user
      operationId: deleteUser
      tags:
        - user
      security:
        - Jwt: []
      responses:
        "204":
          description: Delete the user corresponding to the provided `id`
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
...
  /reservation:
    post:
      summary: Create a reservation
      operationId: createReservation
      tags:
        - reservation
      security:
        - Jwt: []
      responses:
        "201":
          description: The reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationId"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Exists"
        "500":
          $ref: "#/components/responses/Unexpected"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationDetails"
...
  /reservation/user/{id}:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      summary: Read the reservations of the user
      operationId: readReservationUser
      tags:
        - reservation
      responses:
        "200":
          description: The reservations corresponding to the provided `userId`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationList"

        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
  /reservation/lab/{id}:
    parameters:
      - $ref: "#/components/parameters/labId"
    get:
      summary: Read the reservations of the lab
      operationId: readReservationLab
      tags:
        - reservation
      responses:
        "200":
          description: The reservations corresponding to the provided `labId`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Unexpected"
components:
  schemas:
    Id:
      description: ObjectID
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      example: "507f1f77bcf86cd799439011"
    Editable:
      type: boolean
...
    Interval:
      description: Time interval
      type: object
      required:
        - start
        - end
      properties:
        start:
          type: string
          format: date-time
          example: "1970-01-01T00:00:00.000Z"
        end:
          type: string
          format: date-time
          example: "1970-01-01T00:00:00.000Z"
    SlotId:
      type: string
      minLength: 1
    UserId:
      description: The unique identifier of a user
      $ref: "#/components/schemas/Id"
      example: 12412201
    LabId:
      description: The unique identifier of a lab
      $ref: "#/components/schemas/Id"
      example: 1
    ReservationId:
      description: The unique identifier of a reservation
      $ref: "#/components/schemas/Id"
      example: 1
    UserName:
      type: object
      required:
        - first
        - last
      properties:
        first:
          type: string
          example: "Juan"
        last:
          type: string
          example: "Dela Cruz"
    UserPassword:
      type: string
      format: password
      example: "123456"
      writeOnly: true
    UserLogin:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          $ref: "#/components/schemas/UserPassword"
    UserDetails:
      type: object
      required:
        - name
        - description
      properties:
        name:
          $ref: "#/components/schemas/UserName"
        description:
          type: string
          example: "Epic gamer"
...
    ReservationUpdate:
      type: object
      required:
        - schedule
        - slotIds
      properties:
        anonymous:
          type: boolean
        schedule:
          $ref: "#/components/schemas/Interval"
        slotIds:
          type: array
          items:
            $ref: "#/components/schemas/SlotId"
    ReservationDetails:
      allOf:
        - $ref: "#/components/schemas/ReservationUpdate"
        - type: object
          required:
            - labId
          properties:
            labId:
              $ref: "#/components/schemas/LabId"
    ReservationList:
      type: array
      items:
        allOf:
          - $ref: "#/components/schemas/ReservationDetails"
          - type: object
            required:
              - id
            properties:
              id:
                $ref: "#/components/schemas/ReservationId"
              userId:
                $ref: "#/components/schemas/UserId"
              editable:
                $ref: "#/components/schemas/Editable"
    Error:
      type: object
      required:
        - message
      properties:
        message:
          description: A human readable error message
          type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: No items found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Exists:
      description: Data already Exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unexpected:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  parameters:
    userId:
      name: id
      description: The unique identifier of the user
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserId"
    labId:
      name: id
      description: The unique identifier of the lab
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/LabId"
  securitySchemes:
    Jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT